Add prototype new RevTreeBuilder infrastructure.
Factors out RevTreeBuilder as an interface.

Renamed the old RevTreeBuilder as LegacyTreeBuilder and kept it as the default.
The new one can be enabled with the treebuilder.new=true System property (e.g.
export JAVA_OPTS=treebuilder.new=true when using the CLI).

The new infrastructure is in the org.locationtech.geogig.model.experimental
package and decouples the building process from the decision of where in
a tree a given node lies (which bucket index at each tree depth) by means
of the ClusteringStrategy interface. The canonical tree structure remains being
mandated by the old NodePathStorageOrder, which has been renamed as
CanonicalNodeNameOrder to better reflect it's purpose.

Creating a tree with a different structure is, hence, merely a matter of
implementing a different ClusteringStrategy. For example, to create a tree
that represents a Quad-tree Index, a ClusteringStrategy that assigns nodes
to one of the four quadrants based on the node's bounding box can be implemented,
or to create a tree that's an index on a String feature property, one that
sorts based on the feature node's name.

Note however, the new tree builder is memory bound for the time being. So is
LegacyTreeBuilder, but it's main drawback is it buffers Node updates in memory
and flushes to the ObjectStore every one million nodes. Now, the problem with
that is the number of dangling RevTree objects left in the database when the
number of Node updates exceeds that number, meaning every one million nodes
a whole new tree (and it's bucket trees) will be dumped to the databse in order
to preserve Heap memory, and all but the latest will actually be used.

The new tree builder, instead, relies in a mutable data structure that can
be updated in-place until the final RevTree is to be built. The package
org.locationtech.geogig.model.experimental.internal currently hosts the
implementation details to create the mutable tree, abstract out it's
temporary storage, and create the final RevTree out of the mutable representation.

As an example of it's efficiency, creating a 10 milltion nodes RevTree to
an in-memory ObjectStore takes:
                        Time    #of RevTrees stored     Storage size
LegacyTreeBuilder   : ~3.4 min    758,198             755,876,725 bytes
CanonicalTreeBuilder: ~50 secs    268,614             164,594,256 bytes

Signed-off-by: Gabriel Roldan <groldan@boundlessgeo.com>